+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project are documented in this file.
 
+## X.X.X
+
+- `status.conditions` missing error handled
+
 ## 0.5.0
 
 - Exception handling during the Auto Reload